backend/requirements.txt
# 核心框架
fastapi==0.117.1
uvicorn[standard]==0.36.0
python-multipart==0.0.20

# 数据库
sqlalchemy==2.0.43
alembic==1.16.5

# 认证和安全
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4

# 数据验证
pydantic==2.11.9
pydantic-settings==2.10.0

# AI框架
langchain==1.0.0a6
langchain-openai==1.0.0a2
langchain-core==1.0.0a4

# WebSocket
websockets==15.0.1

# 工具库
python-dotenv==1.1.1
httpx==0.28.1
aiofiles==23.2.1
